Finished first cut at counting vocabulary cards.
For 'vocabulaire' cards, we count the number of notes that specify the
'revers' field. (We ignore notes with a blank 'revers' field because
those are likely extra cards beyond the base cards.)

For 'vocabulaire - cloze' cards, we divide the number of notes by 2,
since each word has approximately 2 notes associated with it.