Word grouping script: extract cloze words as well as standard words.
Originally I wasn't dealing with the words inside of vocabulaire - cloze
notes, because I didn't want to futz with them, but I found that
limiting. A lot of standard vocabulaire notes are nouns because they are
easiest to represent on that note type. But verbs, adjectives, and
adverbs more frequently show up on vocabulaire - cloze notes.

This change makes the groupings much more interesting and less noun
heavy.